refactor: clean up unused code and improve issue reference utilities
