+++ pyproject.toml
@@ -61,4 +61,6 @@ dev = [
     "hypothesis>=6.151.9",
     "pytest-asyncio>=0.21.0",
     "httpx>=0.24.0",
+    "radon>=6.0.1",
+    "vulture>=2.16",
 ]
