+++ src/vibe3/exceptions/__init__.py
@@ -113,50 +113,6 @@ class SerenaError(SystemError):
         self.operation = operation
 
 
-class SQLiteError(SystemError):
-    """SQLite database operation failed."""
-
-    def __init__(self, operation: str, details: str) -> None:
-        """Initialize SQLiteError.
-
-        Args:
-            operation: Database operation that failed
-            details: Error details
-        """
-        super().__init__(f"SQLite {operation} failed: {details}")
-        self.operation = operation
-
-
-class CommitAnalyzerError(SystemError):
-    """CommitAnalyzer git command failed."""
-
-    def __init__(self, operation: str, details: str) -> None:
-        """Initialize CommitAnalyzerError.
-
-        Args:
-            operation: CommitAnalyzer operation that failed
-            details: Error details
-        """
-        super().__init__(f"CommitAnalyzer {operation} failed: {details}")
-        self.operation = operation
-        self.details = details
-
-
-class HookManagerError(SystemError):
-    """HookManager hook install/uninstall failed."""
-
-    def __init__(self, operation: str, details: str) -> None:
-        """Initialize HookManagerError.
-
-        Args:
-            operation: HookManager operation that failed (e.g., 'install', 'uninstall')
-            details: Error details
-        """
-        super().__init__(f"HookManager {operation} failed: {details}")
-        self.operation = operation
-        self.details = details
-
-
 # ========== Business Errors ==========
 
 
@@ -171,63 +127,6 @@ class PRNotFoundError(VibeError):
         """
         super().__init__(f"PR #{pr_number} not found", recoverable=False)
         self.pr_number = pr_number
-
-
-class FlowNotFoundError(VibeError):
-    """Flow does not exist."""
-
-    def __init__(self, flow_slug: str) -> None:
-        """Initialize FlowNotFoundError.
-
-        Args:
-            flow_slug: Flow slug that was not found
-        """
-        super().__init__(f"Flow '{flow_slug}' not found", recoverable=False)
-        self.flow_slug = flow_slug
-
-
-class IssueNotFoundError(VibeError):
-    """Issue does not exist."""
-
-    def __init__(self, issue_number: int) -> None:
-        """Initialize IssueNotFoundError.
-
-        Args:
-            issue_number: Issue number that was not found
-        """
-        super().__init__(f"Issue #{issue_number} not found", recoverable=False)
-        self.issue_number = issue_number
-
-
-class TaskNotFoundError(VibeError):
-    """Task does not exist."""
-
-    def __init__(self, task_id: str) -> None:
-        """Initialize TaskNotFoundError.
-
-        Args:
-            task_id: Task ID that was not found
-        """
-        super().__init__(f"Task '{task_id}' not found", recoverable=False)
-        self.task_id = task_id
-
-
-# ========== Batch Errors ==========
-
-
-class BatchError(SystemError):
-    """Batch operation partially failed."""
-
-    def __init__(self, message: str, errors: list[dict]) -> None:
-        """Initialize BatchError.
-
-        Args:
-            message: Summary error message
-            errors: List of error details, each with
-                at least 'file'/'task' and 'error' keys
-        """
-        super().__init__(f"{message} ({len(errors)} failures)")
-        self.errors = errors
 
 
 # ========== Orchestration Errors ==========
