+++ src/vibe3/services/handoff_service.py
@@ -186,72 +186,6 @@ class HandoffService:
         logger.bind(path=str(handoff_path)).success("Appended handoff update")
         return handoff_path
 
-    def _record_item(
-        self,
-        kind: str,
-        ref: str,
-        next_step: str | None,
-        blocked_by: str | None,
-        actor: str | None,
-        session_id: str | None = None,
-    ) -> None:
-        """Helper to record a handoff item."""
-        branch = self.git_client.get_current_branch()
-        effective_actor = SignatureService.resolve_for_branch(
-            self.store,
-            branch,
-            explicit_actor=actor,
-        )
-        _record_handoff(
-            self.store,
-            self.git_client,
-            kind,
-            ref,
-            next_step,
-            blocked_by,
-            effective_actor,
-            session_id=session_id,
-        )
-
-    def record_plan(
-        self,
-        plan_ref: str,
-        next_step: str | None,
-        blocked_by: str | None,
-        actor: str | None,
-        session_id: str | None = None,
-    ) -> None:
-        """Record plan handoff."""
-        self._record_item(
-            "plan", plan_ref, next_step, blocked_by, actor, session_id=session_id
-        )
-
-    def record_report(
-        self,
-        report_ref: str,
-        next_step: str | None,
-        blocked_by: str | None,
-        actor: str | None,
-        session_id: str | None = None,
-    ) -> None:
-        """Record report handoff."""
-        self._record_item(
-            "report", report_ref, next_step, blocked_by, actor, session_id=session_id
-        )
-
-    def record_audit(
-        self,
-        audit_ref: str,
-        next_step: str | None,
-        blocked_by: str | None,
-        actor: str | None,
-        session_id: str | None = None,
-    ) -> None:
-        """Record audit handoff."""
-        self._record_item(
-            "audit", audit_ref, next_step, blocked_by, actor, session_id=session_id
-        )
-
     def _get_handoff_template(self) -> str:
         """Get minimal handoff template.
 
@@ -308,72 +242,3 @@ def _get_handoff_template(branch: str) -> str:
 
 <!-- Append-only lightweight updates -->
 """
-
-
-# ---------------------------------------------------------------------------
-# Handoff recording (from handoff_recorder.py)
-# ---------------------------------------------------------------------------
-
-
-def _record_handoff(
-    store: SQLiteClient,
-    git_client: GitClient,
-    handoff_type: str,
-    ref: str,
-    next_step: str | None,
-    blocked_by: str | None,
-    actor: str,
-    session_id: str | None = None,
-) -> None:
-    """Record handoff to store."""
-    logger.bind(
-        domain="handoff",
-        action=f"record_{handoff_type}",
-        ref=ref,
-        actor=actor,
-        session_id=session_id,
-    ).info(f"Recording {handoff_type} handoff")
-
-    branch = git_client.get_current_branch()
-
-    # Determine actor role and session role based on handoff type
-    if handoff_type == "plan":
-        actor_role = "planner_actor"
-        session_role = "planner_session_id"
-    elif handoff_type == "report":
-        actor_role = "executor_actor"
-        session_role = "executor_session_id"
-    else:  # audit
-        actor_role = "reviewer_actor"
-        session_role = "reviewer_session_id"
-
-    # Build update kwargs
-    update_kwargs: dict[str, object] = {
-        actor_role: actor,
-        "next_step": next_step,
-        "blocked_by": blocked_by,
-    }
-
-    if session_id:
-        update_kwargs[session_role] = session_id
-
-    # Set the appropriate ref field
-    if handoff_type == "plan":
-        update_kwargs["plan_ref"] = ref
-    elif handoff_type == "report":
-        update_kwargs["report_ref"] = ref
-    else:
-        update_kwargs["audit_ref"] = ref
-
-    # Update flow state
-    store.update_flow_state(branch, **update_kwargs)
-
-    # Add event
-    store.add_event(
-        branch,
-        f"handoff_{handoff_type}",
-        actor,
-        f"{handoff_type.capitalize()} recorded: {ref}",
-    )
-
-    logger.success(f"{handoff_type.capitalize()} handoff recorded")
